split number out into a module
